+++ packages/core/src/js/feedback/FeedbackWidget.tsx
@@ -261,6 +261,8 @@ export class FeedbackWidget extends React.Component<FeedbackWidgetProps, Feedbac
     const imagePickerConfiguration: ImagePickerConfiguration = this.props;
     const text: FeedbackTextConfiguration = this.props;
     const styles: FeedbackWidgetStyles = { ...defaultStyles(theme), ...this.props.styles };
+    const autoCorrect = config.autoCorrect !== false;
+    const spellCheck = config.spellCheck !== false;
     const onCancel = (): void => {
       if (onFormClose) {
         onFormClose();
@@ -310,6 +312,8 @@ export class FeedbackWidget extends React.Component<FeedbackWidgetProps, Feedbac
                 placeholder={text.namePlaceholder}
                 value={name}
                 onChangeText={value => this.setState({ name: value })}
+                autoCorrect={false}
+                spellCheck={false}
               />
             </>
           )}
@@ -325,6 +329,9 @@ export class FeedbackWidget extends React.Component<FeedbackWidgetProps, Feedbac
                 testID="sentry-feedback-email-input"
                 placeholder={text.emailPlaceholder}
                 keyboardType={'email-address' as KeyboardTypeOptions}
+                autoCapitalize="none"
+                autoCorrect={false}
+                spellCheck={false}
                 value={email}
                 onChangeText={value => this.setState({ email: value })}
               />
@@ -342,6 +349,8 @@ export class FeedbackWidget extends React.Component<FeedbackWidgetProps, Feedbac
             value={description}
             onChangeText={value => this.setState({ description: value })}
             multiline
+            autoCorrect={autoCorrect}
+            spellCheck={spellCheck}
           />
           {(config.enableScreenshot || imagePickerConfiguration.imagePicker || this._hasScreenshot()) && (
             <View style={styles.screenshotContainer}>